refactor derive to retain generic functions
